% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_Riemann.R
\name{dist_Riemann}
\alias{dist_Riemann}
\title{Riemannian distance}
\usage{
dist_Riemann(M1, M2, eps = 1e-09)
}
\arguments{
\item{M1}{a SPD Matrix}

\item{M2}{a SPD matrix (same dims of \code{M1}).}

\item{eps}{smallest eigenvalues. The count of eigenvalues smaller than \code{eps} is reported (just a quality control).}
}
\description{
Computes the Riemannian distance using the Fisher (affine-invariant) metric between two Symmetric Positive Definite (SPD) matrices.
The distance is the length of the geodesic connecting \code{M1} to \code{M2}.
}
\references{
\href{Barachant A, Bonnet S, Congedo M, Jutten C (2012) Multi-Class Brain Computer Interface Classification by Riemannian Geometry
IEEE Transactions on Biomedical Engineering 59(4), 920-928.}{https://hal.archives-ouvertes.fr/hal-00681328}
 (see equation 4)
}
\author{
Livio Finos, Marco Congedo
}
