% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_Riemann.R
\name{geodesic_Riemann}
\alias{geodesic_Riemann}
\alias{mean_geometric_Riemann}
\title{Riemannian geodesic}
\usage{
geodesic_Riemann(M1, M2, t)
}
\arguments{
\item{M1}{a SPD Matrix}

\item{M2}{a SPD matrix (same dims of \code{M1}).}

\item{t}{(real number) arc-length of the geodesic from \code{M1} to \code{M2} 
(e.g., t=0 yields \code{M1}, t=1 yields \code{M2}, 
t=.5 yields the geometric mean of \code{M1} and \code{M2}).}
}
\description{
Computes the Riemannian geodesic using the Fisher (affine-invariant) metric between two Symmetric Positive Definite (SPD) matrices.
\code{mean_geometric_Riemann} is the geodesic with \code{t=.5}.
}
\references{
\href{Barachant A, Bonnet S, Congedo M, Jutten C (2012) Multi-Class Brain Computer Interface Classification by Riemannian Geometry
IEEE Transactions on Biomedical Engineering 59(4), 920-928.}{https://hal.archives-ouvertes.fr/hal-00681328}
 (see equation 4)
}
\author{
Livio Finos, Marco Congedo
}
