% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_Riemann.R
\name{norm_Riemann}
\alias{norm_Riemann}
\title{Riemannian norm}
\usage{
norm_Riemann(M1, eps = 1e-09)
}
\arguments{
\item{M1}{a SPD Matrix}

\item{eps}{smallest eigenvalues. The count of eigenvalues smaller than \code{eps} is reported (just a quality control).}
}
\description{
computes the Riemannian norm of a SPD matrix; 
it is defined as the distance (\code{dist_Riemann}) between \code{M1} and the identity matrix.
}
\references{
\href{Barachant A, Bonnet S, Congedo M, Jutten C (2012) Multi-Class Brain Computer Interface Classification by Riemannian Geometry
IEEE Transactions on Biomedical Engineering 59(4), 920-928.}{https://hal.archives-ouvertes.fr/hal-00681328}
 (see equation 4)
}
\author{
Livio Finos, Marco Congedo
}
