% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-Riemann.R
\docType{package}
\name{Package Introduction}
\alias{Package Introduction}
\alias{Riemann-package}
\title{Riemann package}
\description{
We provide a variety of algorithms for manifold-valued data, including Frechet summaries, 
hypothesis testing, clustering, visualization, and other learning tasks.
}
