% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_GlobalGeary.R
\name{riem.geary}
\alias{riem.geary}
\title{Geary's \emph{C}}
\usage{
riem.geary(riemobj, W, alternative = c("two.sided", "greater", "less"), ...)
}
\description{
Geary's \emph{C}
}
\concept{spatial}
