% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_GlobalMoran.R
\name{riem.moran}
\alias{riem.moran}
\title{Moran's \emph{I}}
\usage{
riem.moran(riemobj, W, alternative = c("two.sided", "greater", "less"), ...)
}
\description{
Moran's \emph{I}
}
\concept{spatial}
