% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-Riemann.R
\docType{package}
\name{Riemann-package}
\alias{Riemann}
\alias{Riemann-package}
\title{Riemann package}
\description{
We provide a variety of algorithms for manifold-valued data, including Frechet summaries,
hypothesis testing, clustering, visualization, and other learning tasks.
}
\section{License}{

Author-owned software in the September 21, 2026 revision of version 0.2.0 is
licensed under the GNU General Public License, version 3 (GPL-3).
Earlier MIT releases and preserved archives retain their original licenses.
Third-party notices are in the installed \file{COPYRIGHTS} file.
Dataset attribution and any data-specific terms are separate; consult the
dataset help pages. The separate manuscript and replication workspace has
its own license notices.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://www.kisungyou.com/Riemann/}
  \item Report bugs at \url{https://github.com/kisungyou/Riemann/issues}
}

}
\author{
\strong{Maintainer}: Kisung You \email{kisung.you@outlook.com} (\href{https://orcid.org/0000-0002-8584-459X}{ORCID})

Authors:
\itemize{
  \item Kisung You \email{kisung.you@outlook.com} (\href{https://orcid.org/0000-0002-8584-459X}{ORCID})
}

}
