% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Riemann.R
\docType{package}
\name{Riemann}
\alias{Riemann}
\alias{Riemann-package}
\title{Riemann: A package to compute the Riemannian distance using the Fisher (affine-invariant) metric between two Symmetric Positive Definite (SPD) matrices ando other useful tools.}
\description{
Riemann: A package to compute the Riemannian distance using the Fisher (affine-invariant) metric between two Symmetric Positive Definite (SPD) matrices ando other useful tools.
}
\section{Riemann functions}{

The Riemann functions ...
}

