% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual_sphere_mixnorm.R
\name{sphere.mixnorm}
\alias{sphere.mixnorm}
\title{Finite Mixture of Spherical Normal Distributions}
\usage{
sphere.mixnorm(
  datalist,
  k = 2,
  same.lambda = FALSE,
  variants = c("soft", "hard", "stochastic"),
  ...
)
}
\description{
Finite Mixture of Spherical Normal Distributions
}
