% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_gorilla.R
\docType{data}
\name{gorilla}
\alias{gorilla}
\title{Gorilla skull data}
\format{
a named list containing\describe{
\item{male}{a 3d array of size \eqn{(8\times 2\times 29)}}
\item{female}{a 3d array of size \eqn{(8\times 2\times 30)}}
}
}
\usage{
data(gorilla)
}
\description{
This is 29 male and 30 female gorillas' skull landmark data where each 
individual is represented as 8-ad/landmarks in 2 dimensions. This is a 
re-arranged version of the data from \pkg{shapes} package.
}
\examples{
# LOAD AND WRAP AS RIEMOBJ
data(gorilla)
riem.female = wrap.landmark(gorilla$female)
riem.male   = wrap.landmark(gorilla$male)

}
\references{
\insertRef{dryden_statistical_2016}{Riemann}

\insertRef{reno_sexual_2003}{Riemann}
}
\seealso{
\code{\link{wrap.landmark}}
}
\concept{data}
\keyword{datasets}
