% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture_methods.R
\name{loglkd}
\alias{loglkd}
\title{Compute the log-likelihood}
\usage{
loglkd(object, newdata)
}
\arguments{
\item{object}{a fitted mixture model of \code{riemmix} class.}

\item{newdata}{data of \eqn{n} objects (vectors, matrices) that can be wrapped by one of \code{wrap.*} functions in the \pkg{Riemann} package.}
}
\value{
the log-likelihood.
}
\description{
Given a fitted mixture model \eqn{f(x)} and observations \eqn{x_1, \ldots, x_n \in \calM}, compute the log-likelihood
\deqn{L = \log \prod_{i=1}^n f(x_i) = \sum_{i=1}^n \log f(x_i)}.
}
