% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future_special_spd_pdist.R
\name{spd.pdist}
\alias{spd.pdist}
\title{Pairwise Distance on SPD Manifold}
\usage{
spd.pdist(riemobj, geometry, as.dist = FALSE)
}
\value{
a S3 \code{dist} object or \eqn{(N\times N)} symmetric matrix of pairwise distances according to \code{as.dist} parameter.
}
\description{
Pairwise Distance on SPD Manifold
}
\concept{spd}
\keyword{internal}
