% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_splaplace.R
\name{splaplace}
\alias{splaplace}
\alias{dsplaplace}
\alias{rsplaplace}
\alias{mle.splaplace}
\title{Spherical Laplace Distribution}
\usage{
dsplaplace(data, mu, sigma, log = FALSE)

rsplaplace(n, mu, sigma)

mle.splaplace(data, method = c("Newton", "Halley", "Optimize", "DE"), ...)
}
\arguments{
\item{mu}{a length-\eqn{p} unit-norm vector of location.}

\item{sigma}{a scale parameter that is positive.}

\item{n}{the number of samples to be generated.}
}
\description{
Spherical Laplace Distribution
}
\concept{distribution}
