% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_spnorm.R
\name{spnorm}
\alias{spnorm}
\alias{dspnorm}
\alias{rspnorm}
\alias{mle.spnorm}
\title{Spherical Normal Distribution}
\usage{
dspnorm(datalist, mu, lambda, log = FALSE)

rspnorm(n, mu, lambda)

mle.spnorm(datalist, method = c("Newton", "Halley", "Optimize", "DE"), ...)
}
\description{
Spherical normal (SN) distribution
}
\concept{distribution}
